Make all reviews block share the same frontend script
