+++ assets/js/blocks/reviews/frontend.js
@@ -6,10 +6,10 @@ import { render } from 'react-dom';
 /**
  * Internal dependencies
  */
-import FrontendContainerBlock from '../frontend-container-block.js';
+import FrontendContainerBlock from './frontend-container-block.js';
 
 const containers = document.querySelectorAll(
-	'.wp-block-woocommerce-reviews-by-product'
+	'.wp-block-woocommerce-reviews-by-product, .wp-block-woocommerce-reviews-by-category'
 );
 
 if ( containers.length ) {
@@ -22,6 +22,7 @@ if ( containers.length ) {
 			showReviewImage: el.classList.contains( 'has-image' ),
 			showReviewRating: el.classList.contains( 'has-rating' ),
 			showReviewContent: el.classList.contains( 'has-content' ),
+			showProductName: el.classList.contains( 'has-product-name' ),
 		};
 
 		render( <FrontendContainerBlock attributes={ attributes } />, el );
