+++ assets/js/blocks/reviews/reviews-by-category/edit.js
@@ -3,7 +3,6 @@
  */
 import { __, _n, sprintf } from '@wordpress/i18n';
 import {
-	BlockControls,
 	InspectorControls,
 } from '@wordpress/editor';
 import {
@@ -11,44 +10,24 @@ import {
 	PanelBody,
 	Placeholder,
 	ToggleControl,
-	Toolbar,
 	withSpokenMessages,
 } from '@wordpress/components';
 import { SearchListItem } from '@woocommerce/components';
 import { Fragment } from '@wordpress/element';
-import { compose } from '@wordpress/compose';
 import PropTypes from 'prop-types';
-import { debounce } from 'lodash';
 
 /**
  * Internal dependencies
  */
-import EditorBlock from './editor-block.js';
 import ProductCategoryControl from '../../../components/product-category-control';
 import { IconReviewsByCategory } from '../../../components/icons';
-import { getSharedReviewContentControls, getSharedReviewListControls } from '../edit.js';
-import { getBlockClassName, getOrderArgs } from '../utils.js';
+import { renderViewMode, getBlockControls, getSharedReviewContentControls, getSharedReviewListControls } from '../edit-utils.js';
 
 /**
  * Component to handle edit mode of "Reviews by Category".
  */
 const ReviewsByCategoryEditor = ( { attributes, debouncedSpeak, setAttributes } ) => {
 	const { editMode, categoryIds, showReviewDate, showReviewerName, showReviewContent, showProductName, showReviewImage, showReviewRating } = attributes;
-
-	const getBlockControls = () => (
-		<BlockControls>
-			<Toolbar
-				controls={ [
-					{
-						icon: 'edit',
-						title: __( 'Edit' ),
-						onClick: () => setAttributes( { editMode: ! editMode } ),
-						isActive: editMode,
-					},
-				] }
-			/>
-		</BlockControls>
-	);
 
 	const renderCategoryControlItem = ( args ) => {
 		const { item, search, depth = 0 } = args;
@@ -163,37 +142,20 @@ const ReviewsByCategoryEditor = ( { attributes, debouncedSpeak, setAttributes } 
 		);
 	};
 
-	const renderViewMode = () => {
-		if ( ! showReviewContent && ! showReviewRating && ! showReviewDate && ! showReviewerName && ! showReviewImage && ! showProductName ) {
-			return renderHiddenContentPlaceholder();
-		}
-
-		const { reviewsOnPageLoad } = attributes;
-		const { order, orderby } = getOrderArgs( attributes.orderby );
-
-		return (
-			<div className={ getBlockClassName( 'wc-block-reviews-by-category', attributes ) }>
-				<EditorBlock
-					attributes={ attributes }
-					categoryIds={ categoryIds }
-					delayFunction={ ( callback ) => debounce( callback, 400 ) }
-					orderby={ orderby }
-					order={ order }
-					reviewsToDisplay={ reviewsOnPageLoad }
-				/>
-			</div>
-		);
-	};
-
 	if ( ! categoryIds || editMode ) {
 		return renderEditMode();
 	}
 
+	const isAllContentHidden = ! showReviewContent && ! showReviewRating && ! showReviewDate && ! showReviewerName && ! showReviewImage && ! showProductName;
+
 	return (
 		<Fragment>
-			{ getBlockControls() }
+			{ getBlockControls( editMode, setAttributes ) }
 			{ getInspectorControls() }
-			{ renderViewMode() }
+			{ isAllContentHidden ?
+				renderHiddenContentPlaceholder() :
+				renderViewMode( 'wc-block-reviews-by-category', attributes )
+			}
 		</Fragment>
 	);
 };
@@ -215,6 +177,4 @@ ReviewsByCategoryEditor.propTypes = {
 	debouncedSpeak: PropTypes.func.isRequired,
 };
 
-export default compose( [
-	withSpokenMessages,
-] )( ReviewsByCategoryEditor );
+export default withSpokenMessages( ReviewsByCategoryEditor );
