+++ assets/js/blocks/reviews/reviews-by-category/index.js
@@ -11,7 +11,7 @@ import '../editor.scss';
 import Editor from './edit';
 import { IconReviewsByCategory } from '../../../components/icons';
 import sharedAttributes from '../attributes';
-import { getBlockClassName } from '../utils.js';
+import save from '../save.js';
 
 /**
  * Register and run the "Reviews by category" block.
@@ -55,21 +55,5 @@ registerBlockType( 'woocommerce/reviews-by-category', {
 	/**
 	 * Save the props to post content.
 	 */
-	save( { attributes } ) {
-		const { imageType, orderby, categoryIds, reviewsOnPageLoad, reviewsOnLoadMore, showLoadMore, showOrderby } = attributes;
-
-		const data = {
-			'data-image-type': imageType,
-			'data-category-ids': categoryIds.join( ',' ),
-			'data-orderby': orderby,
-			'data-reviews-on-page-load': reviewsOnPageLoad,
-			'data-reviews-on-load-more': reviewsOnLoadMore,
-			'data-show-load-more': showLoadMore,
-			'data-show-orderby': showOrderby,
-		};
-
-		return (
-			<div className={ getBlockClassName( 'wc-block-reviews-by-category', attributes ) } { ...data } />
-		);
-	},
+	save,
 } );
