+++ assets/js/blocks/reviews/reviews-by-product/edit.js
@@ -3,50 +3,30 @@
  */
 import { __, _n, sprintf } from '@wordpress/i18n';
 import {
-	BlockControls,
 	InspectorControls,
 } from '@wordpress/editor';
 import {
 	Button,
 	PanelBody,
 	Placeholder,
-	Toolbar,
 	withSpokenMessages,
 } from '@wordpress/components';
 import { SearchListItem } from '@woocommerce/components';
 import { Fragment } from '@wordpress/element';
 import PropTypes from 'prop-types';
-import { debounce } from 'lodash';
 
 /**
  * Internal dependencies
  */
-import EditorBlock from './editor-block.js';
 import ProductControl from '../../../components/product-control';
 import { IconReviewsByProduct } from '../../../components/icons';
-import { getSharedReviewContentControls, getSharedReviewListControls } from '../edit.js';
-import { getBlockClassName, getOrderArgs } from '../utils.js';
+import { renderViewMode, getBlockControls, getSharedReviewContentControls, getSharedReviewListControls } from '../edit-utils.js';
 
 /**
  * Component to handle edit mode of "Reviews by Product".
  */
 const ReviewsByProductEditor = ( { attributes, debouncedSpeak, setAttributes } ) => {
 	const { editMode, productId, showReviewDate, showReviewerName, showReviewContent, showReviewImage, showReviewRating } = attributes;
-
-	const getBlockControls = () => (
-		<BlockControls>
-			<Toolbar
-				controls={ [
-					{
-						icon: 'edit',
-						title: __( 'Edit' ),
-						onClick: () => setAttributes( { editMode: ! editMode } ),
-						isActive: editMode,
-					},
-				] }
-			/>
-		</BlockControls>
-	);
 
 	const renderProductControlItem = ( args ) => {
 		const { item = 0 } = args;
@@ -156,37 +136,20 @@ const ReviewsByProductEditor = ( { attributes, debouncedSpeak, setAttributes } )
 		);
 	};
 
-	const renderViewMode = () => {
-		if ( ! showReviewContent && ! showReviewRating && ! showReviewDate && ! showReviewerName && ! showReviewImage ) {
-			return renderHiddenContentPlaceholder();
-		}
-
-		const { reviewsOnPageLoad } = attributes;
-		const { order, orderby } = getOrderArgs( attributes.orderby );
-
-		return (
-			<div className={ getBlockClassName( 'wc-block-reviews-by-product', attributes ) }>
-				<EditorBlock
-					attributes={ attributes }
-					delayFunction={ ( callback ) => debounce( callback, 400 ) }
-					orderby={ orderby }
-					order={ order }
-					productId={ productId }
-					reviewsToDisplay={ reviewsOnPageLoad }
-				/>
-			</div>
-		);
-	};
-
 	if ( ! productId || editMode ) {
 		return renderEditMode();
 	}
 
+	const isAllContentHidden = ! showReviewContent && ! showReviewRating && ! showReviewDate && ! showReviewerName && ! showReviewImage;
+
 	return (
 		<Fragment>
-			{ getBlockControls() }
+			{ getBlockControls( editMode, setAttributes ) }
 			{ getInspectorControls() }
-			{ renderViewMode() }
+			{ isAllContentHidden ?
+				renderHiddenContentPlaceholder() :
+				renderViewMode( 'wc-block-reviews-by-product', attributes )
+			}
 		</Fragment>
 	);
 };
