+++ assets/js/blocks/reviews/reviews-by-product/index.js
@@ -11,7 +11,7 @@ import '../editor.scss';
 import Editor from './edit';
 import { IconReviewsByProduct } from '../../../components/icons';
 import sharedAttributes from '../attributes';
-import { getBlockClassName } from '../utils.js';
+import save from '../save.js';
 
 /**
  * Register and run the "Reviews by Product" block.
@@ -47,21 +47,5 @@ registerBlockType( 'woocommerce/reviews-by-product', {
 	/**
 	 * Save the props to post content.
 	 */
-	save( { attributes } ) {
-		const { imageType, orderby, productId, reviewsOnPageLoad, reviewsOnLoadMore, showLoadMore, showOrderby } = attributes;
-
-		const data = {
-			'data-image-type': imageType,
-			'data-product-id': productId,
-			'data-orderby': orderby,
-			'data-reviews-on-page-load': reviewsOnPageLoad,
-			'data-reviews-on-load-more': reviewsOnLoadMore,
-			'data-show-load-more': showLoadMore,
-			'data-show-orderby': showOrderby,
-		};
-
-		return (
-			<div className={ getBlockClassName( 'wc-block-reviews-by-product', attributes ) } { ...data } />
-		);
-	},
+	save,
 } );
