+++ src/BlockTypes/ReviewsByProduct.php
@@ -44,7 +44,7 @@ class ReviewsByProduct extends AbstractBlock {
 	 * @return string Rendered block type output.
 	 */
 	public function render( $attributes = array(), $content = '' ) {
-		\Automattic\WooCommerce\Blocks\Assets::register_block_script( $this->block_name . '-frontend' );
+		\Automattic\WooCommerce\Blocks\Assets::register_block_script( 'reviews-frontend' );
 
 		return $content;
 	}
